Hide the `form` element if the users disable JavaScript
This commit adds a `hidden` attribute to the `form` element by default. It only shows up if the users allow JavaScript to run on their browsers.

The purpose is to prevent the users submit empty email input when
they disable JavaScript since there is not `required` attribute in the
input.